rating pop up worked on

+++ app/Http/Controllers/SettingsController.php
@@ -3,23 +3,35 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use App\Models\headside;
 use App\Models\AboutUs;
 use App\Models\FooterContacts;
+use App\Models\SubscriberSection;
 
 class SettingsController extends Controller
 {
     public function index(){
+        $user = Auth::user();
+        if(!$user){
+           return view('auth.login');
+        }
         $headside = headside::all()->take('1');
         $aboutus = AboutUs::all()->take('1');
+        $subscribersection = SubscriberSection::all()->take('1');
         $footercontacts = FooterContacts::all();
         return view('pages.mysettings')
         ->with('headside',$headside)
         ->with('aboutus',$aboutus)
+        ->with('subscribersection',$subscribersection)
         ->with('footercontacts',$footercontacts);
     }
 
     public function ChangeSideHeadText(Request $req){
+        $user = Auth::user();
+        if(!$user){
+           return view('auth.login');
+        }
          $text = headside::find(1);
          $req->validate(['title'=>'required',
          'body'=>'required',
@@ -36,6 +48,10 @@ class SettingsController extends Controller
     }
 
     public function ChangeAboutUs(Request $req){
+        $user = Auth::user();
+        if(!$user){
+           return view('auth.login');
+        }
         $text = AboutUs::find(1);
         $req->validate([
         'body'=>'required',
@@ -48,6 +64,10 @@ class SettingsController extends Controller
    }
 
    public function updatefc(Request $req,$id){
+         $user = Auth::user();
+         if(!$user){
+            return view('auth.login');
+         }
          $footerC = FooterContacts::find($id);
          $req->validate([
          'phone1'=>'required',
@@ -61,7 +81,24 @@ class SettingsController extends Controller
           ->with('success','Changes to footer have been made');
         }
 
+    public function updatesubscribersection(Request $req){
+        $user = Auth::user();
+        if(!$user){
+        return view('auth.login');
+         }
+        $subscriber_msg = SubscriberSection::find(1);
+         $subscriber_msg->email_msg = $req->input("emailmessage");
+         $subscriber_msg->mobile_msg = $req->input("mobilemessage");
+         $subscriber_msg->save();
+         return back()
+         ->with('success','Changes to footer have been made');
+       }
+
     public function addFootterContent(Request $req){
+        $user = Auth::user();
+        if(!$user){
+           return view('auth.login');
+        }
          $orders = new FooterContacts;
          $orders->phone1 = $req->input("phone1");
          $orders->phone2 = $req->input("phone2");
@@ -73,6 +110,10 @@ class SettingsController extends Controller
     }
 
     public function deleteFc($id){
+        $user = Auth::user();
+        if(!$user){
+           return view('auth.login');
+        }
         $orders = FooterContacts::find($id);
         $orders->delete();
         return back()
